fully implemented image sorting by brightness
-stripped away color sorting and moved it to its own branch

-toggle for sort order

-reload images button

-proper integration with storagevalues

-proper null/invalid folder
checks/wanrings

__

-added switchedTo() for sorters to notify when they are selected via GUI

-slighty refactored how resetting storagevalues can be restricted

-renamed shrinking methods in Util

-refactored and simplified how
fixedsizevisualizers maximize components

+++ src/main/java/main/visualizers/NumberFixedVisualizer.java
@@ -1,7 +1,5 @@
 package main.visualizers;
 
-import java.awt.*;
-
 import main.VisualSortingTool;
 import main.algorithms.Algorithm;
 import main.sorters.Sorter.Sorters;
@@ -12,6 +10,8 @@ import main.ui.custimization.values.StorageValue;
 import main.util.StringHelper;
 import main.vcs.VisualComponent;
 import main.visualizers.bases.FixedSizeVisualizer;
+
+import java.awt.*;
 
 public class NumberFixedVisualizer extends FixedSizeVisualizer
 {
@@ -46,7 +46,7 @@ public class NumberFixedVisualizer extends FixedSizeVisualizer
 	{
 		g.setColor(confirmed ? Algorithm.confirmationColor : highlights[i]);
 		g.setFont(new Font("", Font.PLAIN, componentSize/2));
-		String value = "" + array[i].getValue();
+		String value = "" + (int) array[i].getValue();
 		//draws number centered around the passed in coors
 		StringHelper.drawCenteredString(value, x+componentSize/2, y+componentSize/2, g);
 	}
